code refactor and file rename
