+++ src/newtons_method.py
